transformers/appairflow7/requirements.txt
apache-airflow==2.3.0
azure-storage-file-share==12.4.1
azure-storage-blob==12.8.1
pandas==1.3.3
pyarrow==5.0.0
